fix(Table): change base HOC useage
